Added PolygonVertexData
Each Vertex (Vert) now has associated with it (indirectly) one or more
data packets (for example texture coordinates)that are related to both
itself and a polygon that intend to use that vertex in its rendering.

The Vertexes get loaded with the relevant vertex data (held in a
PolygonVertexData object) from the polygon just before that polygon
renders itself.